requirements.txt
llama-index
langchain
gradio
openai
python-dotenv